Refactor Height class and add operator+ overload
This commit refactors the Height class by adding a friend function to overload the + operator. The operator+ function takes two Height objects and returns their sum. The centimeters are added and converted to meters if necessary. The commit also includes a main function that demonstrates the usage of the operator+ overload.

Closes #123